#2 FirebaseAuthのエラーハンドリングをFirebaseAuthクラスのExtensionに変更
