+++ Tumolink2iOSClient/Controller/LoginVC.swift
@@ -44,7 +44,7 @@ class LoginVC: UIViewController {
         Auth.auth().signIn(withEmail: email, password: password) { (user, error) in
             if let error = error {
                 debugPrint(error.localizedDescription)
-                self.handleFireAuthError(error: error)
+                Auth.auth().handleFireAuthError(error: error, vc: self)
                 self.activityIndicator.stopAnimating()
                 return
             }
