+++ Tumolink2iOSClient/XIBs/ForgotPasswordVC.swift
@@ -33,7 +33,7 @@ class ForgotPasswordVC: UIViewController {
         Auth.auth().sendPasswordReset(withEmail: email) { (error) in
             if let error = error {
                 debugPrint(error.localizedDescription)
-                self.handleFireAuthError(error: error)
+                Auth.auth().handleFireAuthError(error: error, vc: self)
                 return
             }
